% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflames.R
\name{inflames_read}
\alias{inflames_read}
\title{Read 'Queimadas DB' file as SF object.}
\usage{
inflames_read(ifile)
}
\arguments{
\item{ifile}{Path to Queimadas DB CSV file}
}
\value{
Raster with the BNDVI data.
}
\description{
This function read data from 'Queimadas DB / INPE' as \link[sf:sf]{sf::st_sf} object.
}
